Move SelectableData in archive package
