+++ babble/src/main/java/io/mosaicnetworks/babble/fragments/discover/DiscoverGroupsFragment.java
@@ -46,8 +46,9 @@ import io.mosaicnetworks.babble.node.ConfigManager;
 import io.mosaicnetworks.babble.servicediscovery.ResolvedGroup;
 
 /**
- * This fragment implements a visual component to allow selection of a discovered group. It includes
- * swipe refresh and a nothing found message
+ * This fragment implements a visual component to allow selection of a discovered group. It
+ * aggregates groups discovered via mDNS or WebRTC discovery. It includes swipe refresh and a
+ * nothing found message.
  */
 public class DiscoverGroupsFragment extends Fragment {
